Fix: long 타입에서 BigDeCimal로 변경 (#215)
* fix: int -> long

* fix: oauth2 도메인 설정

* refactor: 로그 한 줄 추가

* fix: 도메인 문제....

* test: 테스트 제거

* refactor: AOP Proceed 제거

* fix : long -> BigDeCimal

* test: 테스트 수정

+++ src/main/java/com/jelly/zzirit/domain/order/mapper/OrderMapper.java
@@ -33,7 +33,7 @@ public class OrderMapper {
 		return Order.builder()
 			.member(member)
 			.orderNumber(orderNumber)
-			.totalPrice(BigDecimal.valueOf(dto.totalAmount()))
+			.totalPrice(dto.totalAmount())
 			.status(OrderStatus.PENDING)
 			.shippingRequest(dto.shippingRequest())
 			.address(dto.address())
